rename package name to forest
